Optimize image loading, update link styles, and improve right sidebar layout
