+++ src/app/blog/[category]/[subject]/[title]/[content]/page.tsx
@@ -1,77 +1,40 @@
-import fs from "fs";
-import matter from "gray-matter";
-import { dbtable } from "@/config/types";
-import { components } from "@/components/mdx";
-import { MDXRemote } from "next-mdx-remote/rsc";
-import remarkGfm from "remark-gfm";
-import remarkBreaks from "remark-breaks";
-import rehypePrettyCode from "rehype-pretty-code";
-import rehypeKatex from "rehype-katex";
-import "katex/dist/katex.min.css";
-import remarkMath from "remark-math";
-import rehypeSlug from "rehype-slug";
+import { urlParams } from "@/config/types";
 import RightSidebarComp from "@/components/bars/RightSidebar";
-import rehypeAutolinkHeadings from "rehype-autolink-headings";
 import { CategoryKo } from "@/config/koname";
 import { Pw } from "@/components/Pw";
 import { Suspense } from "react";
 import { Toup } from "@/components/buttons/Toup";
 import { Todown } from "@/components/buttons/Todown";
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbList,
-  BreadcrumbPage,
-  BreadcrumbSeparator,
-} from "@/components/ui/breadcrumb";
-export interface Post extends dbtable {
-  url: string;
-  slug: string;
-  categoryPath: string;
-  content: string;
-  categoryPublicName: string;
-}
-
-interface params {
-  category: string;
-  subject: string;
-  title: string;
-  content: string;
-}
+import { Metadata } from "next";
+import { BreadCrumb } from "@/components/posts/BreadCrumb";
+import { MdxBody } from "@/components/posts/MdxBody";
+import { MdxHeader } from "@/components/posts/MdxHeader";
+import { getPostDetail } from "@/lib/PostUtils/GetPost";
 
 type Props = {
-  params: params;
+  params: urlParams;
+  searchParams: { [key: string]: string | string[] | undefined };
 };
 
-const getPostDetail = async (params: params) => {
-  const filePath = `${process.cwd()}/src/posts/${params.category}/${params.subject}/${params.title}/${params.content}.mdx`;
-  const detail = await parsePost(filePath);
-  return detail;
-};
-const parsePostDetail = async (postPath: string) => {
-  console.log(postPath);
-  const file = fs.readFileSync(postPath, "utf8");
-  const { data, content } = matter(file);
-  const grayMatter = data as dbtable;
-  return { ...grayMatter, content };
-};
-const parsePost = async (postPath: string): Promise<any> => {
-  const postDetail = await parsePostDetail(postPath);
+export async function generateMetadata(
+  { params }: Props,
+  parent: any,
+): Promise<Metadata> {
+  // read route params
+  const id = params.title;
+  // fetch data
+  // const product = await fetch(`https://.../${id}`).then((res) => res.json());
+  // optionally access and extend (rather than replace) parent metadata
+  const previousImages = (await parent).openGraph?.images || [];
   return {
-    ...postDetail,
+    title: id,
+    openGraph: {
+      images: ["/some-specific-page-image.jpg", ...previousImages],
+    },
   };
-};
+}
 
-// For MySQL... Maybe Later...
-// const getMarkdownsource = async () => {
-//   const markdownsource: dbtable[] = await fetch(`${process.env.API_URL}/api`, {
-//     cache: "no-store",
-//   }).then((res) => res.json());
-//   return markdownsource;
-// };
-
-const Post = async ({ params }: Props) => {
+const Post = async ({ params }: { params: urlParams }) => {
   const post = await getPostDetail(params);
   return (
     <>
@@ -79,126 +42,12 @@ const Post = async ({ params }: Props) => {
       <Suspense fallback={<div>Loading...</div>}>
         {CategoryKo[params.category].lock && <Pw />}
       </Suspense>
-      <Breadcrumb>
-        <BreadcrumbList>
-          <BreadcrumbItem>
-            <BreadcrumbLink
-              href={`/blog/${params.category}`}
-              className="hover:font-bold hover:text-text"
-            >
-              {CategoryKo[params.category].name.split("(")[0]}
-            </BreadcrumbLink>
-          </BreadcrumbItem>
-          <BreadcrumbSeparator />
-          <BreadcrumbItem>
-            <BreadcrumbLink
-              href={`/blog/${params.category}/${params.subject}`}
-              className="hover:font-bold hover:text-text"
-            >
-              {
-                CategoryKo[params.category].sub[params.subject].name.split(
-                  "(",
-                )[0]
-              }
-            </BreadcrumbLink>
-          </BreadcrumbItem>
-          <BreadcrumbSeparator />{" "}
-          <BreadcrumbItem>
-            <BreadcrumbLink
-              href={`/blog/${params.category}/${params.subject}/${params.title}`}
-              className="hover:font-bold hover:text-text"
-            >
-              {
-                CategoryKo[params.category].sub[params.subject].title[
-                  params.title
-                ].name.split("(")[0]
-              }
-            </BreadcrumbLink>
-          </BreadcrumbItem>
-          <BreadcrumbSeparator />
-          <BreadcrumbItem>
-            <BreadcrumbPage className="font-bold text-text">
-              {
-                CategoryKo[params.category].sub[params.subject].title[
-                  params.title
-                ].content[params.content].name.split("(")[0]
-              }
-            </BreadcrumbPage>
-          </BreadcrumbItem>
-          <div className={"ml-auto text-right"}>
-            {"📅 " +
-              CategoryKo[params.category].sub[params.subject].title[
-                params.title
-              ].content[params.content].date}
-          </div>
-        </BreadcrumbList>
-      </Breadcrumb>
+      <BreadCrumb params={params} />
+      <RightSidebarComp content={post.content} />
       <div className="prose mt-10 min-h-[100vh] scroll-smooth dark:prose-invert focus:scroll-auto">
-        <RightSidebarComp content={post.content} />
         <Suspense fallback={<div>Loading...</div>}>
-          <MDXRemote
-            source={post.content === "" ? "no contents 😿" : post.content}
-            //@ts-ignore
-            components={components}
-            options={{
-              parseFrontmatter: true,
-              mdxOptions: {
-                remarkPlugins: [remarkGfm, remarkBreaks, remarkMath],
-                rehypePlugins: [
-                  [
-                    // 이슈 존재 https://github.com/hashicorp/next-mdx-remote/issues/86
-                    //@ts-ignore
-                    rehypePrettyCode,
-                  ],
-                  [
-                    //@ts-ignore
-                    rehypeKatex,
-                    {
-                      colorIsTextColor: true,
-                      strict: false,
-                      macros: {
-                        // issue from google fonts
-                        "\\neq": "\\mathrel{\\char`≠}",
-                      },
-                    },
-                  ],
-                  rehypeSlug,
-                  [
-                    rehypeAutolinkHeadings,
-                    {
-                      behavior: "append",
-                      properties: {
-                        className: ["anchor"],
-                      },
-                      content: {
-                        type: "element",
-                        tagName: "span",
-                        properties: { className: ["icon", "icon-link"] },
-                        children: [{ type: "text", value: "🔗" }],
-                      },
-                    },
-                  ],
-                ],
-              },
-            }}
-          />
-          {/*For MySQL... Maybe Later...*/}
-          {/*{markdownsource.map((data: dbtable, idx: any) => (*/}
-          {/*  <div key={idx}>*/}
-          {/*    <h1 className={"text-pink-600 hover:text-amber-500"}>*/}
-          {/*      {data.title}*/}
-          {/*    </h1>*/}
-          {/*    <p>{data.description}</p>*/}
-          {/*    <p>{data.date.toString()}</p>*/}
-          {/*    <p>{data.thumbnail}</p>*/}
-          {/*    <MDXRemote*/}
-          {/*      source={data.content}*/}
-          {/*      components={MyComponents}*/}
-          {/*      options={{ parseFrontmatter: true }}*/}
-          {/*    />*/}
-          {/*    <br />*/}
-          {/*  </div>*/}
-          {/*))}*/}
+          <MdxHeader props={post} />
+          <MdxBody content={post.content} />
         </Suspense>
       </div>
       <Todown />
